+++ tests/test_resource_tika.py
@@ -6,7 +6,7 @@ from tika_client.data_models import Document
 
 
 class TestParseFormatted:
-    def test_parse_docx(self, tika_client: TikaClient):
+    def test_parse_docx_from_file_as_html(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.docx"
         resp = tika_client.tika.as_html.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
@@ -14,7 +14,7 @@ class TestParseFormatted:
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert "<body><p>This is an DOCX test document, also made September 14, 2022</p>\n</body>" in resp.content
 
-    def test_parse_docx_no_mime(self, tika_client: TikaClient):
+    def test_parse_docx_no_mime_from_file_as_html(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.docx"
         resp = tika_client.tika.as_html.from_file(test_file)
 
@@ -22,7 +22,7 @@ class TestParseFormatted:
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert "<body><p>This is an DOCX test document, also made September 14, 2022</p>\n</body>" in resp.content
 
-    def test_parse_microsoft_word_docx(self, tika_client: TikaClient):
+    def test_parse_microsoft_word_docx_from_file_as_html(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "microsoft-sample.docx"
         resp = tika_client.tika.as_html.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
@@ -33,7 +33,7 @@ class TestParseFormatted:
             in resp.content
         )
 
-    def test_parse_odt(self, tika_client: TikaClient):
+    def test_parse_odt_from_file_as_html(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.odt"
         resp = tika_client.tika.as_html.from_file(test_file, magic.from_file(str(test_file), mime=True))
 
@@ -43,7 +43,7 @@ class TestParseFormatted:
 
 
 class TestParsePlain:
-    def test_parse_docx(self, tika_client: TikaClient):
+    def test_parse_docx_from_file_as_text(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.docx"
 
         resp = tika_client.tika.as_text.from_file(test_file, magic.from_file(str(test_file), mime=True))
@@ -52,7 +52,7 @@ class TestParsePlain:
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert "This is an DOCX test document, also made September 14, 2022" in resp.content
 
-    def test_parse_odt(self, tika_client: TikaClient):
+    def test_parse_odt_from_file_as_text(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.odt"
 
         resp = tika_client.tika.as_text.from_file(test_file, magic.from_file(str(test_file), mime=True))
@@ -63,7 +63,7 @@ class TestParsePlain:
 
 
 class TestParseContentPlain:
-    def test_parse_docx_bytes(self, tika_client: TikaClient):
+    def test_parse_docx_from_bytes_buffer(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.docx"
         buffer = test_file.read_bytes()
 
@@ -73,7 +73,7 @@ class TestParseContentPlain:
         assert resp.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
         assert "This is an DOCX test document, also made September 14, 2022" in resp.content
 
-    def test_parse_odt_bytes(self, tika_client: TikaClient):
+    def test_parse_odt_from_bytes_buffer(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.odt"
         buffer = test_file.read_bytes()
 
@@ -83,7 +83,17 @@ class TestParseContentPlain:
         assert resp.type == "application/vnd.oasis.opendocument.text"
         assert "This is an ODT test document, created September 14, 2022" in resp.content
 
-    def test_parse_buffer_text_content(self, tika_client: TikaClient):
+    def test_parse_odt_from_bytes_buffer_with_mime(self, tika_client: TikaClient):
+        test_file = SAMPLE_DIR / "sample.odt"
+        buffer = test_file.read_bytes()
+
+        resp = tika_client.tika.as_text.from_buffer(buffer, "application/vnd.oasis.opendocument.text")
+
+        assert isinstance(resp, Document)
+        assert resp.type == "application/vnd.oasis.opendocument.text"
+        assert "This is an ODT test document, created September 14, 2022" in resp.content
+
+    def test_html_document_from_string_buffer(self, tika_client: TikaClient):
         test_file = SAMPLE_DIR / "sample.html"
         buffer = test_file.read_text()
 
@@ -94,3 +104,14 @@ class TestParseContentPlain:
         assert "Hello world! This is HTML5 content in a file for" in resp.data["X-TIKA:content"]
         assert resp.data["dc:title"] == "This Is A Test"
         assert resp.data["description"] == "A sample HTML file"
+
+
+class TestParseContentcompress:
+    def test_parse_odt_from_bytes_buffer_compress(self, tika_client_compressed: TikaClient):
+        test_file = SAMPLE_DIR / "test-document-images.odt"
+        buffer = test_file.read_bytes()
+
+        resp = tika_client_compressed.tika.as_text.from_buffer(buffer)
+
+        assert isinstance(resp, Document)
+        assert resp.type == "application/vnd.oasis.opendocument.text"